Create movietheater.c
Movie Theatre Ticket Pricing with Censor Board Certificate Rating

A movie theatre wants to create a program that calculates ticket prices based on the customer's age, the time of day they are attending the movie, their choice of seating (standard or premium), and the censor board certificate rating of the movie. Write a C program that asks the user for their age (range from 1 to 99), the time of day they are attending the movie (24 hours format), their choice of seating (standard or premium), and the censor board certificate rating of the movie (U or A). The program should then calculate and display the ticket price according to the following rules:

·        If the customer is under 12 years old and the movie has a "U" certificate, the program should ask if they want a children's ticket. If yes, the ticket price is Rs. 200; if no, the regular pricing applies.

·        If the customer is over 12 years old and the movie has a "U" certificate:

·        If they are attending a matinee (before 5 pm), the ticket price is Rs. 300 for standard seating and Rs. 400 for premium seating.

·        If they are attending a show after 5 pm, the ticket price is Rs. 400 for standard seating and Rs. 500 for premium seating.

 if the customer is under 12 years:
 If the movie has an "A" certificate, the program should display a message that the customer is not allowed to watch the movie due to their age.  (Display message is “CUSTOMER IS NOT ALLOWED”).
 If the customer is under 18 years and above 12 years old:
If the movie has an "A" certificate, the program should display a message that the customer can only watch the movie if they are accompanied by an adult. (Display message is “CUSTOMER REQUIRES ADULT COMPANY”) and ticket price will be Rs. 350/-
If the customer's age is 18 or older and the movie has an "A" certificate, the ticket price will be fixed at Rs. 600, regardless of whether they choose standard or premium seating.

Use appropriate conditional statements (if-else) to determine the ticket price based on the age, time of day, seating choice, and censor board certificate rating. Ensure that the program handles invalid inputs and displays error messages when necessary.

Constraints:

·        Age should be integer and the valid range between 1 and 99.

·        Time of the day should be in 24 hours format. (requires only hour part)

·        Choice of seating should be in char data type (‘s’ -> Standard and ‘p’ -> Premium)

·        Censor rating should be in char data type and accepts ‘U’ or ‘A’

Input and Output:

Input age, time_of_the_day, Seating_type, Censor_rating as follows:

Example 1:

25          18          s         U

Output will be:

400

Example 2:

10          18          s           A

Output will be:

CUSTOMER IS NOT ALLOWED